Allow clearing task preference value. Refactored unit tests to separate file.
